update api books

+++ BookstoreApi/BookstoreApi/Services/CategoriesService.cs
@@ -8,9 +8,11 @@ namespace BookstoreApi.Services;
 public class CategoriesService
 {
     private readonly IMongoCollection<Category> _categoryCollection;
+    private readonly BooksService _booksService;    
 
     public CategoriesService(
-        IOptions<BookStoreDatabaseSettings> bookStoreDatabaseSettings)
+        IOptions<BookStoreDatabaseSettings> bookStoreDatabaseSettings,
+        BooksService booksService)
     {
         var mongoClient = new MongoClient(
             bookStoreDatabaseSettings.Value.ConnectionString);
@@ -20,6 +22,7 @@ public class CategoriesService
 
         _categoryCollection = mongoDatabase.GetCollection<Category>(
             bookStoreDatabaseSettings.Value.CategoriesCollectionName);
+        _booksService = booksService;
     }
 
     public async Task<List<Category>> GetAsync()
@@ -27,9 +30,9 @@ public class CategoriesService
         return await _categoryCollection.Find(_ => true).ToListAsync();
     }
 
-    public async Task<Category?> GetCategoryNameAsync(string Category)
+    public async Task<Category?> GetAsync(string Id)
     {
-        return await _categoryCollection.Find(x => x.CategoryName == Category).FirstOrDefaultAsync();
+        return await _categoryCollection.Find(x => x.Id == Id).FirstOrDefaultAsync();
     }
 
     public async Task CreateAsync(Category newCategory)
@@ -40,7 +43,17 @@ public class CategoriesService
     public async Task UpdateAsync(string id, Category updatedCategory) =>
         await _categoryCollection.ReplaceOneAsync(x => x.Id == id, updatedCategory);
 
-    public async Task RemoveAsync(string id) =>
+    public async Task RemoveAsync(string id)
+    {
+        var category = await _categoryCollection.Find(x => x.Id == id).FirstOrDefaultAsync();
+
+        if (category is null)
+        {
+            return;
+        }
         await _categoryCollection.DeleteOneAsync(x => x.Id == id);
+        // Update the CategoryId to null in all books with the deleted category id
+        await _booksService.UpdateCategoryAsync(id, null);        
+    }
 
 }
